add IB and replace ap exam specific state with more generic name
